requirements.txt
sqlalchemy
pymysql
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
